pnordd: disable large reload feature of SFC
The pblog seemed to hit a corner case in the SFCs caching when jumping
between the end of a flash page and the beginning of the previous
flash page resulting in corrupt data being returned repeatedly for one read.

Disabling this provides the correct behavior in all cases but does
result in a few additional seconds of boot time.

(cherry picked from commit 154dea2a26cb8a5fec98293da4f1dc575c3cda32)

Change-Id: Ic6f6b49436d4e4c232416b275b0a35acf39e84ab
Signed-off-by: Robert Lippert <rlippert@google.com>
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/25019
Tested-by: Jenkins Server
Tested-by: Jenkins OP Build CI
Tested-by: Jenkins OP HW
Tested-by: FSP CI Jenkins
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>