Allow ADU access to memory instead of PBA in ecmd debug tool
The mempba access procedures use PBA BAR3 to get to memory.  The
OCC also uses BAR3 so there is a collision with the tool if the OCC
is running.  Added an option to the ecmd framework to allow a user
to force a method (pba/adu) and for the tool to check by default.

Change-Id: I403cf1f08a8f753ea7f29e59501d80fb2daf432f
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/14677
Tested-by: Jenkins Server
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>