eRepair MBVPD size check for CDIMM and ISDIMM
The eRepair VPD size for DMI Bus is bigger in systems with ISDIMMs because
the VPD accounts for all the Centaurs. The current check in the code is
for each Centaur which is applicable for systems with CDIMMs.

The fix also includes a change to do restore of eRepair records noted
in the VPD when the system had a non-fips830 driver.

Change-Id: Ibaaf98c5be3e26474ee0eede6081f506f4bb7ec3
RTC: 119532
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/14727
Tested-by: Jenkins Server
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Reviewed-by: Prem Shanker Jha <premjha2@in.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>