PRD: Removing the Threshold Sync between HB and FSP
This change is part of a broader change to move PRD Mfg thresholds
from being maintained in a file to being maintained as system attributes.
Here, we are removing the code that was used to synchronize the PRD Mfg
thresholds between FSP and Hostboot. We do not need it because the thresholds
which now are common attributes are synchronized outside of PRD context.

Change-Id: I0a738b5f5056ce51a484367b3648b6ccb0448947
RTC: 118150
CMVC-Prereq: SW287619
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/14776
Tested-by: Jenkins Server
Reviewed-by: Prem Shanker Jha <premjha2@in.ibm.com>
Reviewed-by: Sachin Gupta <sgupta2m@in.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>
Reviewed-by: Zane Shelley <zshelle@us.ibm.com>
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/15873