memboot: Add a pnor driver that supports memboot
Memboot is a mechanism where we program the AST2400 BMC LPC->AHB
bridge to access main memory on the BMC instead of flash memory. This
allows the host to be booted from a temporary firmware without
flashing, however it requires the firmware to understand this and not
reinitialise the LPC->AHB bridge.

This patch reads the boot mode SuperIO scratch register and selects
the correct method for accessing the system firmware image.

Signed-off-by: Alistair Popple <alistair@popple.id.au>