Initialize I2C Switches in struct to zero
There are multiple uses for the I2C Switches in the misc_args_t
struct used in I2C operations.  The struct was not defaulting the
switches value to zero and there were negative consequences of this
early in the IPL for the i2cPresence() function.

Fixes open-power/hostboot#18

Change-Id: I83d34770ff04cfe31bf07b320d8821c8fa80c705
CQ: SW2999529
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/16447
Tested-by: Jenkins Server
Reviewed-by: MATTHEW A. PLOETZ <maploetz@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>