i2c: only send 1 master stop in FORCE_UNLOCK reset mode
The existing code path would send 2 master stops, one in
i2cForceResetAndUnlock() and another in i2cSendSlaveStop().

The one in i2cSendSlaveStop() can cause IPL failures if the
VSPD voltage is not online for a centaur I2C port because
it requires checking that the I2C bus is "free" (pulled up)
before reporting success.

Change-Id: I435f4e738d1c000e23167221a42f6747f0dce93b
Signed-off-by: Robert Lippert <rlippert@google.com>
Signed-off-by: William A. Kennington III <wak@google.com>