kernel: consider writable/clean pages as evictable
The PnorRP maps "readonly" PNOR sections as WRITABLE so that
procedures can scribble on the memory if necessary.  But this
causes the kernel to never cast out these (usually) read-only
pages and eventually run out of memory.

To fix this only consider a writable+dirty page as actually
unevictable, and treat a writable+clean page as readonly similarly
to other logic in this file.

Change-Id: I9f822d11f4b285f49f13090f29d11f0858d3487d
Signed-off-by: Robert Lippert <rlippert@google.com>
Signed-off-by: William A. Kennington III <wak@google.com>