Set All I2C Master Engine Mode Registers For PHYP
Previously, Hostboot code was only setting the I2C Master Engine
mode registers if hostboot thought that there was a device off of
at least one of the ports connected to the master/engine.  This
commit now sets all of the mode registers (master engines 0 and 1)
for all of the processors in the system.

Change-Id: I1371fb12f44c675b076d2e094e6513f74aff5856
CQ: SW302964
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/16995
Tested-by: Jenkins Server
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Reviewed-by: WILLIAM G. HOFFA <wghoffa@us.ibm.com>
Reviewed-by: STEPHEN M. CPREK <smcprek@us.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>