Allow I2C Resets on Processor Engine 1
This commit removes a workaround that limited all I2C resets to
just engine 0 on processors. (Centaurs only use engine 0).  Now,
if there is a device off of one of the buses connected to engine 1
(port 0, 1, or 2), then a reset will be issued on the bus.

Change-Id: Id2091b44d931be5321420c091fe91a803cc5df65
CQ: SW300339
Backport: release-fips830
CMVC-Prereq: 955865
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/17309
Tested-by: Jenkins Server
Reviewed-by: STEPHEN M. CPREK <smcprek@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Reviewed-by: WILLIAM G. HOFFA <wghoffa@us.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>