Improve I2C Master Setup By Checking Master State
Before setting up the I2C Masters for PHYP there is now a check that
the engine is in a good state.  If it is in a bad state, an I2C reset
is issued in an attempt to recover the master before setting it up.

Change-Id: Ib023366fc416bb9d5de9fcd0b885c879f22b918f
CQ: SW295172
Backport: release-fips830
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/17524
Tested-by: Jenkins Server
Reviewed-by: STEPHEN M. CPREK <smcprek@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>