Rediscover location of HDAT data to handle IO CM operations
PHYP modifies the location of the HDAT memory during a IO CM
operation.  During a mpipl, during dump collection, HB will
cache the location of the SPIRA.  This cache will be pointing
to the B side.  Later in the boot the FSP repopulates the SPIRA
using the A side.  HB needs to use the A side pointers instead.

Change-Id: I991fc731eb6292869e7280d29727ef270a6975a5
CQ: SW311349
Backport: release-fips830
Backport: release-fips820
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/18836
Tested-by: Jenkins Server
Tested-by: Jenkins OP Build CI
Tested-by: Jenkins OP HW
Reviewed-by: Elizabeth Liner <eliner@us.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>