Read SBE_VITAL_REG from Master Processor for OpenPower Systems
The SBE Update code needs to read the SBE_VITAL_REG to determine which
SBE Seeprom a processor used to boot.  For OpenPower slave processors
this register has not been setup yet at the time it is read, so the
master processor's SBE_VITAL_REG will be used instead.

Change-Id: I22d9c21c44cf9be9cf5da58bb73cd1e0bcb576b0
CQ: SW312152
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/19527
Tested-by: Jenkins Server
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins OP Build CI
Tested-by: Jenkins OP HW
Reviewed-by: Dean Sanner <dsanner@us.ibm.com>
Reviewed-by: MATTHEW A. PLOETZ <maploetz@us.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>