Update SBE Image Version to recognize GOLDEN SBE Image
This commit adds a flag "origin" to the SBE version info at 0x300 of the
SBE to indicate whether the current working side SBE image originated
from the golden side.  When a system has trouble booting from the primary
side PNOR and fails over to the golden side PNOR, hostboot copies the golden
side SBE over to the primary side SBE.  Without the origin flag, there's no
reliable way to know, after the fact, which SBE image is on the primary side.
The flag's value was added to a trace message in getSeepromSideVersion, to
indicate the flag's value to the user.

Change-Id: I1ac3ab6a7239b2302a334bcf1f1bd7f664b7844b
RTC:128646
ForwardPort: yes
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/23617
Tested-by: Jenkins Server
Reviewed-by: WILLIAM G. HOFFA <wghoffa@us.ibm.com>
Tested-by: Jenkins OP Build CI
Tested-by: Jenkins OP HW
Tested-by: FSP CI Jenkins
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>