Race condition in IPMI-SEL thread start
This fixes a defect where we shutdown Hostboot before the SEL that
caused the shutdown is sent to the BMC.

The SEL processing of error logs is done in the background by a thread
that is only started when the first error log is committed.  Due to
lack of proper synchronization, the shutdown could previously occur
before the SEL thread had fully started and registered itself in the
shutdown processing to ensure its queue was flushed.

Modified the IPMI-SEL thread start up to synchronize with a barrier
after it has completed its init actions (such as registering for
shutdown).  This ensures that the thread triggering the start up does
not continue, and trigger shutdown, until the initialization is
complete.

Change-Id: Ie25a5c6e81580375fd18dfe4431660ca14f78273
CQ:SW322712
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/21270
Tested-by: Jenkins Server
Tested-by: Jenkins OP Build CI
Tested-by: Jenkins OP HW
Tested-by: FSP CI Jenkins
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>