PRD: Check for checkstop or unit cs before clearing service call flag
A core checkstop can be blamed on a recoverable error. Some recoverable
errors can be set up to not trigger predictive/visible logs. If a core
checkstop or system checkstop is blamed on one of these, we can't clear
the service call flag. This shouldn't generally be a concern with system
checkstops, since thresholdAndMask bits should also be given a secondary
filter. But if there were a case of thresholdAndMask that is not
secondary, this fix should also cover that.

Change-Id: If4d58d7abbcb5383c6779f4d8f249a93fcb58ee2
CQ: SW324604
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/21287
Tested-by: Jenkins Server
Reviewed-by: Caleb N. Palmer <cnpalmer@us.ibm.com>
Reviewed-by: Brian J. Stegmiller <bjs@us.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>
Reviewed-by: BENJAMIN J. WEISENBECK <bweisenb@us.ibm.com>
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/21414
Tested-by: Jenkins OP Build CI
Tested-by: Jenkins OP HW
Tested-by: FSP CI Jenkins