ECID Invalid in Error Log
ECID values have been showing incorrectly in logs.  The binary
versions of the error logs look OK.  The issue only affects the
ASCII version of the error log.  The code that parses the binary
was found to contain pointer arithmetic with missing parenthesis
resulting in arithmetic to the actual value of the data instead of
the pointer value.  This commit contains changes to xmltohb.pl,
which generates the C++ code for the parser.  Eight such instances
of erroneous pointer arithmetic, including the one that causes the
invalid ECID value, were found and corrected.

Change-Id: Ida63f99df33157ce7dc8c7befe26a6ffa8c9d41f
CQ:321666
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/21394
Tested-by: Jenkins Server
Reviewed-by: WILLIAM G. HOFFA <wghoffa@us.ibm.com>
Tested-by: Jenkins OP Build CI
Tested-by: Jenkins OP HW
Tested-by: FSP CI Jenkins
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>