genlist/gensyms: Remove pthread usage
Usage of pthreads and popen() triggers what I think is a GNU libc bug:
- https://patchwork.ozlabs.org/patch/539329/
- https://sourceware.org/bugzilla/show_bug.cgi?id=19182

I could 100% hit this on Fedora 23, making Hostboot impossible to compile
on modern Fedora.

looking into the code, the pthread usage isn't really needed as since
there's a couple of calls it's going to be done in parallel anyway.

Plus, since there was pretty much no error checking going on, doing
things in parallel is just going to increase the possibility of things
going wrong.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>