SW328412 - Multiple SELs per eSEL
The problem being fixed is the generation of multiple
SELs per eSEL when sending of the eSEL to the BMC
fails due to loss of SEL reservation on the BMC.  The
code in send_esel is not supposed to generate the SEL
if there was an error sending the eSEL(the eSEL is in
the log stored on the PNOR and will be sent after the
next reboot if it has not been marked as having been
sent to the BMC).  The error is that there are two
error variables that need to be checked in order to
determine if there was an error sending the eSEL to
the BMC.  The code was only checking one of those.

Change-Id: I1346b29d8c31ed862b0bcfba9a6b8f057030ac20
CQ:SW328412
backport:master-p8
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/22663
Tested-by: Jenkins Server
Reviewed-by: WILLIAM G. HOFFA <wghoffa@us.ibm.com>
Reviewed-by: Marshall J. Wilks <mjwilks@us.ibm.com>
Reviewed-by: A. Patrick Williams III <iawillia@us.ibm.com>
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/23177
Tested-by: Jenkins OP Build CI
Tested-by: Jenkins OP HW