Fix buildpnor.pl to both work and be slightly more deterministic
The whole process of assembling a pnor image is currently a big
non-deterministic trainwreck and needs a lot more fixing than
this simple patch.

However, in order to start fixing it, we need to make things work
properly to begin with.

With the code that went into 02353bc we seem to have started to rely
on perl hash ordering in order to create the PNOR file successfully.

What was occuring was that we were writing the backup TOCs in an order
that meant we triggered an error in fpart (from open-power/ffs):

fpart: unexpected : ../src/libffs.c(968) : (code=-1) 'BACKUP_PART' at offset 33521664 and size 32768 overlaps 'part' at offset 33521664 and size 4096

This patch enforces order in the universe and (slightly) increases
the determinism of the pnor creation process.

In the near future, I plan to rip out all of this and replace it with a
small amount of pflash invocations as this whole code path is near
incomprehensible nonsense that is probably 10-100x longer than just hard
coding the headers and using dd to inject the contents.

Fixes: 02353bc
Reported-by: Joel Stanley <joel@jms.id.au>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>