Fix buildpnor.pl to both work and be slightly more deterministic
The current code seems to rely on perl hash ordering in order to
create the PNOR file successfully.  What was occuring was that we were
writing the backup TOCs in an order that meant we triggered an error in
fpart (from open-power/ffs):

    > fpart: unexpected : ../src/libffs.c(968) : (code=-1) 'BACKUP_PART'
    > at offset 33521664 and size 32768 overlaps 'part' at offset 33521664
    > and size 4096

This patch enforces order in the universe and increases
the determinism of the pnor creation process.

Change-Id: I3bbaf977542a1bebaf1e0b0b528538a7c54541fa
Reported-by: Joel Stanley <joel@jms.id.au>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>
Signed-off-by: Patrick Williams <iawillia@us.ibm.com>
Forwardport: yes
Github: fixes open-power/hostboot#44
Github: closes open-power/hostboot#45
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/23183
Tested-by: Jenkins Server
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/23187