Fixup HBB address on golden side SBE in the field
This commit adds code to the SBE side actions to perform a one-time HBB fixup
on all machines that have the potential of having the golden side HBB pointer
incorrectly pointing to the primary PNOR side.  This will fix all machines out
in the field on the first boot after upgrading.

Change-Id: I416cdc95f47f15e8254b5b1391a62417470f0728
RTC:331944
ForwardPort: yes
Depends-on:I1ac3ab6a7239b2302a334bcf1f1bd7f664b7844b
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/23618
Tested-by: Jenkins Server
Reviewed-by: WILLIAM G. HOFFA <wghoffa@us.ibm.com>
Tested-by: Jenkins OP Build CI
Tested-by: Jenkins OP HW
Tested-by: FSP CI Jenkins
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>