PRD: disabled references to legacy FAPI interfaces.
Simply commented out the current code and added RTC tags to reference
stories that will restore the functionality when FAPI2 is available.
Also disabled compilation of all the platform plugin code since
that code is the most common user of the legacy FAPI interfaces and
most of it will be rewritten for P9 anyway.

Change-Id: I8cb57c7343e71c83ab9017abfc715198eedbb9f4
RTC: 143877
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/23369
Reviewed-by: Caleb N. Palmer <cnpalmer@us.ibm.com>
Tested-by: Jenkins Server
Reviewed-by: BENJAMIN J. WEISENBECK <bweisenb@us.ibm.com>
Reviewed-by: Zane Shelley <zshelle@us.ibm.com>
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/23676
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>