P9 changes for PR core restriction
Covers the work required to support the P9 module vpd and new EQ/EX/EC core
structure in relation to the processing of the PR keyword to restrict the
number of cores.

Add new traces or enhance old traces with additional data to improve debug.

Update the module vpd binary for our standalone model.

Change-Id: I20018ccf1daa429e5b0ad9e38f50b1bc5943bcb5
RTC:145459
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/705
Tested-by: Jenkins Server
Tested-by: FSP CI Jenkins
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Reviewed-by: Matthew A. Ploetz <maploetz@us.ibm.com>