SW327988: Fix for infinite reboot loop
Removed code that incremented the reboot count sensor when shutting down
due to an error caught outside the reconfigure loop.  With this change, the
BMC will attempt to boot 2 times from the primary side and 1 time from the
golden side, after that the system will shut down and halt.

Change-Id: I60d357568164ab5f5d69fed5db0c65eefa876d42
CQ: SW327988
ForwardPort: yes
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/23951
Tested-by: Jenkins Server
Reviewed-by: MATTHEW A. PLOETZ <maploetz@us.ibm.com>
Reviewed-by: STEPHEN M. CPREK <smcprek@us.ibm.com>
Tested-by: Jenkins OP Build CI
Tested-by: Jenkins OP HW
Tested-by: FSP CI Jenkins
Reviewed-by: WILLIAM G. HOFFA <wghoffa@us.ibm.com>
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/24820