Implement DO_NOT_WAKEUP opMode for Scom access
During runtime we occasionally will read scoms from various chiplets.
Usually if the chiplet is in winkle mode (stopped, idle, etc) then we will
wake it up. If this opmode is set, it is indicating the attempted read is not
very important and not worth the time/energy to wake up the chiplet.

Change-Id: Ieaaffe42fda3459e0295597fcf41c07bbaee72b4
RTC: 150454
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/22986
Tested-by: Jenkins Server
Tested-by: FSP CI Jenkins
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>