Remove the copyright_check that was adding submitter to prolog contributor
The addCopyright script was failing in hostboot-ci because when it was
ran on standalone it would add the current submitter of the commit as a
contributor and when it ran on ci it would add the author of the commit
as a contributor. In the past these two have been the same, however if
we were to merge a pull request from a third-party who was not and IBMer
we would fail CI. Now it will only check the author if the copyright_check
flag is set.

Change-Id: I7a360a824d73a89872040932d2e7e3f8a334e770
RTC:117874
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/24389
Tested-by: Jenkins Server
Tested-by: Jenkins OP Build CI
Tested-by: FSP CI Jenkins
Reviewed-by: Stephen M. Cprek <smcprek@us.ibm.com>
Reviewed-by: Elizabeth K. Liner <eliner@us.ibm.com>
Tested-by: Jenkins OP HW
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>