Add simics breakpoint for exceptions
Created a new magic instruction handler that can be enabled via
environment variable to stop the simulation for Hostboot
exceptions.  This is useful when we want to get a backtrace at
the failing spot without recompiling code or having to manually
insert breakpoints in simics directly.

Enable the function by setting HB_BREAK_ON_ERROR
The trigger call is MAGIC_BREAK_ON_ERROR

Change-Id: I17e008281d010e3f8c5e5817e5f30fd0ccb624d0
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/24600
Tested-by: Jenkins Server
Tested-by: FSP CI Jenkins
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>
Reviewed-by: Marshall J. Wilks <mjwilks@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>