Update ATTR_POSITION_GETMACRO to lookup FAPI_POS instead of POSTION
POSITION is not a reliable way for systems outside Hostboot to track
targets. We will temporarily put this change in to fix an issue we
were having where we were looking up DIMMS based on this attribute
A better solution is to add a MCA_PORT attribute on the DIMM targets
that is defined by the platform.

Change-Id: Ifbb64a9f326de3d08aef488e7a4b0899024a8cbc
RTC: 154950
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/25012
Tested-by: Jenkins Server
Tested-by: FSP CI Jenkins
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Reviewed-by: Matt Derksen <v2cibmd@us.ibm.com>
Reviewed-by: Andres A. Lugo-Reyes <aalugore@us.ibm.com>
Reviewed-by: Matthew A. Ploetz <maploetz@us.ibm.com>