Add explicit delays to centaur sensor cache enable and disable
We discovered that if multiple Centaur i2c operations are running
while the chip is trying to read the dimm temps that we will
eventually time out the OCC due to collisions.  TO fix this we
need to add explicit delays between enable->disable (30ms) and
also between disable->enable (2ms) to allow the sensor cache to
get valid data.

Change-Id: I16c96482f4827097d9af7b1e08dcd42375ae8d30
CQ: SW354148
ForwardPort: yes
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/25723
Tested-by: Jenkins Server
Tested-by: Jenkins OP Build CI
Tested-by: Jenkins OP HW
Tested-by: FSP CI Jenkins
Reviewed-by: Matthew A. Ploetz <maploetz@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/25982