Update FAPI_INVOKE_HWP to reset current_err before falling func
Previously current_err was not getting reset to SUCCESS. This causes
issues when we have expect errors occur (such as test cases) or have
non-fatal errors occur. This commit makes it so that before any
function is called with the FAPI_INVOKE macro, current_err is set
back to SUCCESS so we can be sure are getting valid errors.

Change-Id: I51f137d5b0c30be56e911d5a0e2228db3d609949
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/26010
Tested-by: Jenkins Server
Tested-by: FSP CI Jenkins
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Richard J. Knight <rjknight@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>