Modify default MSS_FREQ used for Centaur SBE init
A system that runs memory at 1600MHz can hit intermittent fails
when we do the Centaur SBE initialization with MSS_FREQ=1066.
The only systems that run at 1600 are CDIMM systems so using that
as the key to decide if we need to apply the code that can run
MSS_FREQ=1066.

Change-Id: Icd3165bb65cc3f1b3320416a510f3dc4e24b98f2
CQ: SW352593
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/27106
Reviewed-by: Thi N. Tran <thi@us.ibm.com>
Reviewed-by: Matthew A. Ploetz <maploetz@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>