Fix pib err mask to actually look at the Userdetails section data
Previously there was a bug with pib_err_mask where we incorrectly
tried to cast a userdetails object into an incorrect structure.
This caused an issue when we tried to access members of the struct
that we were casting to. This commit changes the function that
provides lookup of userdetails sections so it only returns the user
data that the user details section contains.

Change-Id: I38d6bbef8525a177b3f1baf72ea042c8e756c34d
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/28105
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <v2cibmd@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>