Update pnorrp to only attempt to parse TOC if its at a valid offset
In the gensis case for pnor generation they were not creating a backup
toc in pnor. This caused pnorrp to fail because it was attempting to
parse garbage. Our error logs should have caught this but it exposed
a bug in the errlmanager code that tried to call targeting code before
the extended image had been loaded

Change-Id: Iedbb03a9d5a3c842c9497121268d170ab623d075
RTC: 159832
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/28729
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Reviewed-by: Matt Derksen <v2cibmd@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>