Remove explicit calls to p9_cpu_special_wakeup from HWPs
It is illegal for HWPs to call p9_cpu_special_wakeup themselves.
HWPs need to call fapi2::specialWakeup() instead so that the
platform can decide what to do.  There are a couple reasons for this:
1) There may be multiple different pieces of logic running
simultaneously on the platform that are all manipulating the special
wakeup bit in the hardware.  For example, when HBRT is running on top
of PHYP, HBRT can't just mess around with the wakeup bit that PHYP
could already be setting/clearing on its own.
2) The platform owns the recursion counter to handle wakeup being
called while wakeup is already enabled.

Change-Id: I9ff8ac7c6eec566873b1a2f95ce95123982a8fb6
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/28852
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Corey V. Swenson <cswenson@us.ibm.com>
Reviewed-by: Gregory S. Still <stillgs@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>