Check if parent commit is already merged into base branch.
Fixes bug querying commits that were merged on a different branch
When there's a branch split, a commit that was merged on the latest
branch is also on the release branch. There is logic to check if
the parent commit is OPEN or not. Querying for that commit on the
release (relase-fips860) branch fails as Gerrit stored it on the
latest(master-p8).

Change-Id: I62544233069e3513401884066809f2b9cdf8151a
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/29032
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Andrew J. Geissler <andrewg@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Reviewed-by: Thi N. Tran <thi@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>