Remove Memory expansion and TOD hack from call_proc_exit_cache_contained
There are actions going into fips that trigger memory expansion
correctly. This commit removes the previous hacks so we know that
the memory expansion is working properly. NOTE that exit_cache_contained
is still trigged by a SIMICS only register. This is due to issues with
how actions are executed.

Change-Id: I71fec05def230a61996b05c647101518dafb3573
RTC: 158417
CMVC-Prereq: 1004693
CMVC-Prereq: 1005024
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/29033
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>