Remove depreciated PCIE Attributes from p9 nimbus attribute xml
Some of PCIE attributes have changed since p8. HWSV code is still
expecting things that may not be true about these PCIE attrs.
In order to get around this we had to temporarily add _P9 to the end
of the attrs to act as if we were creating new ones. This commit
removes that temporary workaround.

Change-Id: I90dc8efb663fc35ce954d303a6a5dcdf72d91b3b
RTC:150688
CMVC-Prereq: 990341
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/29283
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Tested-by: Daniel M. Crowell <dcrowell@us.ibm.com>