Update ATTR_POWER_CONTROL_REQUESTED enum
The enumeration values changed for this attribute so we need to
update them so that it matches what is in the EKB. As part of this
fix I realized that our xmltohb.pl script that generates C code to
verify ATTR types was not applicable to attrs of type enumeration.
I have updated the script to now ignore enumeration attrs.

RTC:160125
Change-Id: I7836a23600e868a08ae3d1421e1fbb5f43ab9721
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/29349
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Reviewed-by: Matt Derksen <v2cibmd@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>