Do not lose traces during cache to mem transition
There is a window in the simics ipl when we have two objects
covering the same physical addresses.  Our continuous trace
logic needs to pull the data from the same object that is
catching the memory accesses in the model.  This is indicated
by a priority field in the map structure.  (Note - in our
case we want to use the cache object until it goes away in
proc_exit_cache_contained istep.)

Change-Id: I6fc89c142a5fa48b2577867fbf5056ee0bb13289
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/29607
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>