Update fapi2 attribute to have correct persistency
There was a lot of attributes that had incorrect persistency set.
These changes focus on platInit fapi attributes that should have
been set to non-volatile as they are set by the MRW. There are
cases where an attribute will be platInit but not be non-volatile.
This is usually because hostboot needs to access the attribute
or it is an override and should be set to zero between IPLs

Change-Id: I1d73a792564551debe9ca65da1ea2c65b9cc2ae8
RTC: 157411
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/29844
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <v2cibmd@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>