Add REL_POS for units and dimms
The target changes for P9 pointed out the need for a new way
to generically find the position of a unit or dimm at various
levels of the hierarchy.  ATTR_REL_POS represents the position
of a unit/dimm relative to its immediate parent, regardless of
what kind of target the parent is.

Examples:
 proc-0/mcbist-0/mcs-0/mca-1/dimm-2 -> 0 (2 dimms per mca)
 proc-0/mcbist-0/mcs-1/mca-1 -> 1
 proc-0/eq-3/ex-1/core-0 -> 0

Change-Id: I7d7b3de800b167252257a26e7aa03e1eefd66c87
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/30138
Reviewed-by: Zane C. Shelley <zshelle@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: NAGENDRA K. GURRAM <nagendra.g@in.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>